Changed Criterion to use a ReaderT monad with the Config
This stops the Config being passed around as an explicit parameter, which makes the code shorter and cleaner.

I've used the mtl library, but all the ReaderT stuff is wrapped up in the new Criterion.Monad module, so it should be possible to swap the implementation (e.g. for transformers) without any trouble.

One of the main complexities of making this change was to fix hPrintf to work with ReaderT Config IO, rather than IO.  This seems to work, and isn't too horrific.  It actually cleans up the code that uses the Config to decide whether to print something -- that is now nicer.

--HG--
extra : convert_revision : df6ba3f1832d9837ac4c9c32e9c619996696374f