Kill off benchmark comparison as a feature
I never really reviewed this, never use it, and find the code very
obtuse and difficult to understand. This makes it a huge pain to
refactor when I want to add new features.

It seems the more-correct thing to do is to get rid of this feature
and find a better way to implement it that doesn't make the core
of criterion less maintainable.