#107, open the output file in the write mode
With ReadWriteMode Haskell opens the file but doesn't truncate the contents. That means if we have a file there from before, we end up overwriting the start, and if the file was longer, we end up with corrupt JSON. Only visible if you run the test suite multiple times, and somewhat non-deterministic since it requires longer times (when printed in exponential format), followed by shorter times.