Add parserWith to support extensible command-line parsing (#206)
While `parser` only accepts a `Config` as an argument, `parserWith`
accepts an arbitrary `optparse-applicative` parser. This makes it
ideal for using in tandem with custom sets of command-line arguments
that extend the default ones the `criterion` provides.

I've added an example of how one can profitably use `parserWith` in
`examples/ExtensibleCLI.hs`

Fixes #84.