Fix #214 (#215)
If the `nf'`/`whnf'` functions in
`Criterion.Measurement.Types.Internal` aren't sufficiently optimized,
this can cause `criterion`'s innner benchmarking loop to fail to
terminate in certain situations (#214). We hack around this issue
by specifically enabling `-O2 -fno-prof-auto` in that module so that
even if `criterion-measurement` is compiled with `-O0` or
`-fprof-auto`, it won't affect `nf'` or `whnf'`. It's a bit
heavy-handed, but it gets the job done.

Along the way, I decided to add a Travis CI configuration for when
`criterion` and `criterion-measurement` are compiled with `+fast`,
which implies `-O0`. Adding a configuration that builds everything
with `-fprof-auto` seems a bit excessive, so I didn't pursue that.

Fixes #214.