Warn if report names contains newlines
Newlines wreak havoc on JavaScript strings in the default HTML
report. We could attempt to turn newlines into properly escaped JS
strings before compiling the Mustache template, but this is overkill,
since the string will be rendered as if the newlines were simple
spaces anyway. As a result, we simply turn all newlines into spaces
and emit a warning.

Fixes #224.