Escape JSON to avoid parsing the content as HTML (#232)
By crafting a malicious string it was quite easy to escape the HTML
attribute that contained the JSON report data.

This commit aims to mitigate the breakage of the reports when report
names contain unexpected characters. The JSON data is moved to a
<script> tag and the content is escaped by replacing significant HTML
characters with their corresponding JSON escape sequences. The '<'
character is replaced with the sequence "\u003c", '&' is replaced with
"\u0026" and so on.