Use default value for `regressions` when creating `config` (#245)
Previously, the `config` function would ignore the value of
`regressions` that specified (e.g., via the `Config` argument of
`defaultMain`). Since `--regressions` is an argument that is allowed
to be passed multiple times, it's somewhat awkward to specify a default
argument for it. I adopt a similar approach to what the old `arguments1`
function in `optparse-applicative` did (see
https://github.com/pcapriotti/optparse-applicative/commit/700e07b0751d02fec938af67d1f917e60974a601).

Fixes #244.