Make nf' and whnf' robust against eta-expansion changes
As seen in ghc#21948 (https://gitlab.haskell.org/ghc/ghc/-/issues/21948), whether
or not `go` gets eta-expanded crucially determines whether `nf'` is a
function which computes an `IO ()` action or a function which computes a
result of type `(# RealWorld State#, () #)`.

In the former case, the call to `seq` in `runBenchmark` will force the
computation of the `IO ()` action, which can be computed without
dependency on the State# token, and therefore before any IO takes place. This
defeats the point of the function which is supposed to run the loop
inside the IO action (so that measure can correct measure the runtime
stats).

The fix is to make both branches of the loop depend on the state token so
that the loop can't be evaluated before the whole IO action is executed.